// src/modules/department/translate.enum.ts
export enum DepartmentKeys {
    NOT_FOUND = "department.NOT_FOUND",
    CREATION_ERROR = "department.CREATION_ERROR",
    UPDATE_ERROR = "department.UPDATE_ERROR",
    UPDATED_SUCCESSFULLY = "department.UPDATED_SUCCESSFULLY",
    CANT_DELETE = "department.CANT_DELETE",
    DELETED_SUCCESSFULLY = "department.DELETED_SUCCESSFULLY",
    COMPANY_ACCESS_DENIED = "department.COMPANY_ACCESS_DENIED",
    DEPARTMENT_HAS_EMPLOYEES_CANT_DELETE = "department.DEPARTMENT_HAS_EMPLOYEES_CANT_DELETE",
}